requirements.txt
customtkinter
Pillow
google-generativeai
pynput